Updated HelpCommand.
