+++ src/commands/misc/HelpCommand.js
@@ -1,0 +1,45 @@
+const { Command } = require('@greencoast/discord.js-extended');
+const { MessageEmbed } = require('discord.js');
+const { MESSAGE_EMBED } = require('../../common/constants');
+
+class HelpCommand extends Command {
+  constructor(client) {
+    super(client, {
+      name: 'help',
+      aliases: ['h'],
+      description: 'Get a description of all the commands you can use.',
+      emoji: ':question:',
+      group: 'misc',
+      guildOnly: false
+    });
+  }
+
+  prepareFields() {
+    return this.client.registry.groups.map((group) => {
+      const listOfCommands = group.commands.reduce((text, command) => {
+        return text.concat(`${command.emoji} **${this.client.prefix}${command.name}** - ${command.description}\n`);
+      }, '');
+
+      return { title: group.name, text: listOfCommands };
+    });
+  }
+
+  run(message) {
+    const fields = this.prepareFields();
+    const embed = new MessageEmbed()
+      .setTitle('Free Games Notifier Help Message')
+      .setColor(MESSAGE_EMBED.color)
+      .setThumbnail(MESSAGE_EMBED.thumbnail);
+
+    for (const key in fields) {
+      const field = fields[key];
+      embed.addField(field.title, field.text);
+    }
+
+    embed.addField('Found a bug?', `This bot is far from perfect, so in case you found a bug, please report it [here](${MESSAGE_EMBED.issuesURL}).`);
+
+    return message.channel.send(embed);
+  }
+}
+
+module.exports = HelpCommand;
